requirement.txt
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
chardet==5.2.0
cssselect2==0.7.0
defusedxml==0.7.1
lxml==5.1.1
pillow==10.2.0
pycairo==1.27.0
pycparser==2.22
reportlab==4.1.0
svg.py==1.4.3
svglib==1.5.1
tinycss2==1.2.1
webencodings==0.5.1
